requirements.txt
Flask==3.0.0
pandas==2.1.0
scikit-learn==1.3.0
numpy==1.24.3
python-Levenshtein==0.21.1
fuzzywuzzy==0.18.0
matplotlib==3.7.1
seaborn==0.12.2
openai==1.3.0
python-dotenv==1.0.0
networkx==3.1.0
